Moved map editor to tsx
