+++ game-editor/frontend/package.json.md5
@@ -1,1 +1,1 @@
-64e3dda913451a6812749cd019ee8be9
+306aa18d36d2581823df5b4fc90fca32
