+++ game-editor/frontend/src/map-editor/main.tsx
@@ -4,9 +4,7 @@ const MapEditor:React.FC = () => {
     const navigate = useNavigate();
     return(
         <>
-        <button onClick={() => navigate("new-map")}>New Map</button>
-        <br/>
-        <button onClick={() => navigate(-1)}>Go back to main menu</button>
+        <button onClick={() => navigate("new-map")} className="file: bg-blueWhale rounded border-1 border-solid w-1/6 border-black">New Map</button>
         </>
     )
 }
