+++ game-editor/frontend/src/pokemon-editor/main.tsx
@@ -3,76 +3,6 @@ import { ParsePokemonData } from "../../wailsjs/go/main/App";
 import { useEffect, useState } from "react";
 import { Pokemon } from "./pokemon.model";
 
-// interface Pokemon {
-//     ID: string;
-//     Species: string;
-//     Types: string[];
-//     DexEntry: string;
-//     Abilities: {
-//         Name: string;
-//         isHidden: boolean;
-//     }[];
-//     Moves: {
-//         Name: string;
-//         Level: number;
-//         Method: string;
-//     }[];
-//     Evolutions: {
-//         Name: string;
-//         Methods: string[];
-//         ID: string;
-//     }[];
-//     Stats: {
-//         Hp: number;
-//         Attack: number;
-//         Defense: number;
-//         SpecialAttack: number;
-//         SpecialDefense: number;
-//         Speed: number;
-//     };
-//     AssetData: {
-//         Front: string;
-//         Back: string;
-//         ShinyFront: string;
-//         ShinyBack: string;
-//         Icon: string;
-//     };
-// }
-
-// interface Abilities {
-//     Name: string;
-//     isHidden: boolean;
-// }
-
-// interface Moves {
-//     Name: string;
-//     Level: number;
-//     Method: string;
-// }
-
-// interface Evolutions {
-//     Name: string;
-//     Methods: string[];
-//     ID: string;
-// }
-
-// interface Stats {
-//     Hp: number;
-//     Attack: number;
-//     Defense: number;
-//     SpecialAttack: number;
-//     SpecialDefense: number;
-//     Speed: number;
-// }
-
-// interface AssetData {
-//     Front: string;
-//     Back: string;
-//     ShinyFront: string;
-//     ShinyBack: string;
-//     Icon: string;
-// }
-
 export default function PokemonEditor():React.ReactElement {
     const navigate = useNavigate();
     const [pokemonSpecies, setPokemonSpecies] = useState<Pokemon[]>([]);
